Move _base.py to the _pyodide module
