+++ src/py/_pyodide/_core_docs.py
