+++ src/py/pyodide/__init__.py
@@ -1,4 +1,10 @@
-from ._base import open_url, eval_code, eval_code_async, find_imports, CodeRunner
+from _pyodide._base import (
+    open_url,
+    eval_code,
+    eval_code_async,
+    find_imports,
+    CodeRunner,
+)
 from ._core import JsProxy, JsException, create_once_callable, create_proxy, to_js  # type: ignore
 from ._importhooks import jsfinder
 from .webloop import WebLoopPolicy
